All comments can be moderated (reported or not)

+++ view/backend/commentsReportedView.php
@@ -9,12 +9,37 @@
 $title='Modérer les commentaires signalés';
 
 ob_start();
+?>
+<h1>Commentaires signalés</h1>
+<?php
 
-while($comment=$reported->fetch())
+if($countreported>0)
 {
-    echo 'Posté par '.$comment['author'].' le '.$comment['datepost_fr'].' : "'.$comment['content'].'" <a href="index.php?validcomment='.$comment['id'].'">Accepter ce commentaire</a> | <a href="index.php?deletecomment='.$comment['id'].'">Supprimer ce commentaire</a><br>';
+    while($comment=$reported->fetch())
+    {
+        echo 'Posté par '.$comment['author'].' le '.$comment['datepost_fr'].' : "'.$comment['content'].'" <a href="index.php?validcomment='.$comment['id'].'">Accepter ce commentaire</a> | <a href="index.php?deletecomment='.$comment['id'].'">Supprimer ce commentaire</a><br>';
 
+    }
 }
+else{
+    echo 'Aucun commentaires signalés.';
+}
+?>
+
+<h1>Commentaires non signalés</h1>
+<?php
+if($countunreported>0)
+{
+    while($comment=$unreported->fetch())
+    {
+        echo 'Posté par '.$comment['author'].' le '.$comment['datepost_fr'].' : "'.$comment['content'].'" <a href="index.php?deletecomment='.$comment['id'].'">Supprimer ce commentaire</a><br>';
+
+    }
+}
+else{
+    echo 'Aucun commentaires.';
+}
+
 
 $content=ob_get_clean();
 require('view/template.php');
